requirements.txt
fastapi
gunicorn
uvicorn
aiohttp
bs4
pillow
telegraph